move all run config creation code into RunConfigGenerator
