+++ model_analyzer/model_manager.py
@@ -15,7 +15,6 @@
 from model_analyzer.constants import LOGGER_NAME
 from model_analyzer.perf_analyzer.perf_config import PerfAnalyzerConfig
 from model_analyzer.output.file_writer import FileWriter
-from model_analyzer.config.run.run_search import RunSearch
 from model_analyzer.config.run.run_config_generator \
     import RunConfigGenerator
 
@@ -56,10 +55,7 @@ class ModelManager:
         self._metrics_manager = metrics_manager
         self._result_manager = result_manager
         self._state_manager = state_manager
-        self._run_search = RunSearch(config=config)
         self._first_config_variant = None
-        self._run_config_generator = RunConfigGenerator(config=config,
-                                                        client=self._client)
 
         # Generate the output model repository path folder.
         self._output_model_repo_path = config.output_model_repository_path
@@ -75,9 +71,6 @@ class ModelManager:
             The model being run
         """
 
-        # Clear any configs from previous model run
-        self._run_config_generator.clear_configs()
-
         # Reset first config variant for new model
         self._first_config_variant = None
 
@@ -85,157 +78,51 @@ class ModelManager:
         server_config_copy = self._server.config().copy()
         self._server.update_config(params=model.triton_server_flags())
 
-        # Run model inferencing
-        if self._config.run_config_search_disable:
-            logger.info(
-                f"Running manual config search for model: {model.model_name()}")
-            self._run_model_no_search(model)
-        else:
-            logger.info(
-                f"Running auto config search for model: {model.model_name()}")
-            self._run_model_with_search(model)
+        rcg = RunConfigGenerator(config=self._config, client=self._client)
+        rcg.init(model)
+        while not rcg.is_done() and not self._state_manager.exiting():
+            run_config = rcg.next_config()
+            measurement = self._execute_run_config(run_config)
+            rcg.add_measurement(measurement)
 
         # Reset the server args to global config
         self._server.update_config(params=server_config_copy.server_args())
 
-    def _run_model_no_search(self, model):
+    def _execute_run_config(self, run_config):
         """
-        Creates run configs from specified combinations and executes
-        them without any run search
-        """
-
-        # Generate all the run configs at once and return
-        if self._config.triton_launch_mode != 'remote':
-            user_model_config_sweeps = \
-                self._run_config_generator.generate_model_config_combinations(
-                    model.model_config_parameters())
-            for user_model_config_sweep in user_model_config_sweeps:
-                self._run_config_generator.generate_run_config_for_model_sweep(
-                    model, user_model_config_sweep)
-        else:
-            self._run_config_generator.generate_run_config_for_model_sweep(
-                model, None)
-        self._execute_run_configs()
-
-    def _run_model_with_search(self, model):
-        """
-        Searches over the required config elements,
-        creates run configs and executes them
+        Executes the run config. Returns obtained measurement. Also sends 
+        measurement to the result manager
         """
 
-        model_config_parameters = model.model_config_parameters()
+        # Create model variant
+        self._create_model_variant(original_name=run_config.model_name(),
+                                   variant_config=run_config.model_config())
 
-        # Run config search is enabled, figure out which parameters to sweep over and do sweep
-        if self._config.triton_launch_mode == 'remote':
-            self._run_model_config_sweep(model, search_model_config=False)
-        else:
-            user_model_config_sweeps = \
-                self._run_config_generator.generate_model_config_combinations(
-                    model_config_parameters)
-            if model.parameters()['concurrency']:
-                # Both are specified, search over neither
-                for user_model_config_sweep in user_model_config_sweeps:
-                    self._run_config_generator.generate_run_config_for_model_sweep(
-                        model, user_model_config_sweep)
-                self._execute_run_configs()
-            else:
-                # Search through concurrency values only
-                for user_model_config_sweep in user_model_config_sweeps:
-                    if self._state_manager.exiting():
-                        return
-                    self._run_model_config_sweep(
-                        model,
-                        search_model_config=False,
-                        user_model_config_sweep=user_model_config_sweep)
+        # If this run config was already run, do not run again, just get the measurement
+        measurement = self._get_measurement_if_config_duplicate(run_config)
+        if measurement:
+            return measurement
 
-            # If no model config parameters were specified, then we only ran the default
-            # configuration above and need to do an automatic sweep
-            #
-            if not model_config_parameters:
-                self._run_model_config_sweep(model, search_model_config=True)
-
-    def _run_model_config_sweep(self,
-                                model,
-                                search_model_config,
-                                user_model_config_sweep=None):
-        """
-        Initializes the model sweep, iterates until search bounds,
-        and executes run configs
-        """
-
-        self._run_search.init_model_sweep(model.parameters()['concurrency'],
-                                          search_model_config)
-
-        next_model = model
-        while not self._state_manager.exiting():
-
-            # Get next model sweep
-            next_model, auto_model_config_sweep = self._run_search.get_next_model_sweep(
-                next_model)
-
-            # End search when get_model sweep returns empty
-            if not auto_model_config_sweep:
-                break
-            if user_model_config_sweep:
-                model_sweep_for_run_config = user_model_config_sweep
-            else:
-                model_sweep_for_run_config = auto_model_config_sweep[0]
-
-            self._run_config_generator.generate_run_config_for_model_sweep(
-                next_model, model_sweep_for_run_config)
-            self._run_search.add_measurements(self._execute_run_configs())
-
-    def _execute_run_configs(self):
-        """
-        Executes the run configs stored in the run
-        config generator until there are none left.
-        Returns obtained measurements. Also sends them
-        to the result manager
-        """
-
-        measurements = []
-
-        while self._run_config_generator.run_configs():
-            # Check if exiting
-            if self._state_manager.exiting():
-                return measurements
-
-            # Remove one run config from the list
-            run_config = self._run_config_generator.next_config()
-
-            # Create model variant
-            self._create_model_variant(original_name=run_config.model_name(),
-                                       variant_config=run_config.model_config())
-
-            # If this run config was already run, do not run again, just get the measurement
-            measurement = self._get_measurement_if_config_duplicate(run_config)
-            if measurement:
-                measurements.append(measurement)
-                continue
-
-            # Start server, and load model variant
-            self._server.start(env=run_config.triton_environment())
-            if not self._load_model_variant(
-                    variant_config=run_config.model_config()):
-                self._server.stop()
-                continue
-
-            # Profile various batch size and concurrency values.
-            # TODO: Need to sort the values for batch size and concurrency
-            # for correct measurment of the GPU memory metrics.
-            perf_output_writer = None if \
-                not self._config.perf_output else FileWriter(self._config.perf_output_path)
-            perf_config = run_config.perf_config()
-
-            logger.info(f"Profiling model {perf_config['model-name']}...")
-            measurement = self._metrics_manager.profile_model(
-                run_config=run_config, perf_output_writer=perf_output_writer)
-            if measurement is not None:
-                measurements.append(measurement)
-
+        # Start server, and load model variant
+        self._server.start(env=run_config.triton_environment())
+        if not self._load_model_variant(
+                variant_config=run_config.model_config()):
             self._server.stop()
+            return
 
-        return measurements
+        # Profile various batch size and concurrency values.
+        # TODO: Need to sort the values for batch size and concurrency
+        # for correct measurment of the GPU memory metrics.
+        perf_output_writer = None if \
+            not self._config.perf_output else FileWriter(self._config.perf_output_path)
+        perf_config = run_config.perf_config()
+
+        logger.info(f"Profiling model {perf_config['model-name']}...")
+        measurement = self._metrics_manager.profile_model(
+            run_config=run_config, perf_output_writer=perf_output_writer)
+        self._server.stop()
+
+        return measurement
 
     def _create_model_variant(self, original_name, variant_config):
         """
